backend/pyproject.toml
[project]
name = "netflux-backend"
version = "0.1.0"
description = "Backend API for Netflux application"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "sqlmodel>=0.0.16",
    "alembic>=1.13.0",
    "pydantic>=2.5.0",
    "python-multipart>=0.0.6",
    "openai>=1.0.0",
    "aiofiles>=23.2.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
]
